refactor: rename module dir to modules
