+++ packages/sdk/src/__tests__/integration/connectorStructure.spec.ts
@@ -1,5 +1,5 @@
 import { initSDK } from "../../bootstrap";
-import { buildModule } from "../../module/buildModule";
+import { buildModule } from "../../modules/buildModule";
 import { createModuleMock } from "../__mocks__/createModuleMock";
 
 describe("[ConnectorStructure]", () => {
