+++ packages/sdk/src/__tests__/unit/buildModule.spec.ts
@@ -1,6 +1,6 @@
 import { extension1Mock } from "../__mocks__/extension1.mock";
 import module1Mock from "../__mocks__/module1.mock";
-import { buildModule } from "../../module/buildModule";
+import { buildModule } from "../../modules/buildModule";
 
 describe("[buildModule]", () => {
   it("should build module without extensions", () => {
