+++ packages/sdk/src/index.ts
@@ -1,6 +1,6 @@
 import { initSDK } from "./bootstrap";
 import { eventManager } from "./events/EventManager";
-import { buildModule } from "./module/buildModule";
+import { buildModule } from "./modules/buildModule";
 import { handleError } from "./error";
 
 /**
